Renamed files as per new naming.
