+++ src/root.sagas.ts
@@ -1,8 +1,8 @@
 import { SagaMiddleware } from 'redux-saga';
 
-import { uvPieSaga } from './uv_pie/uv_pie-saga';
-import { uvBarChartSaga } from './uv_bar_chart/uv_bar_chart-saga';
-import { uvHeaderSaga } from './uv_header/uv_header-saga';
+import { uvPieSaga } from './uv_pie/uv_pie.saga';
+import { uvBarChartSaga } from './uv_bar-chart/uv_bar-chart.saga';
+import { uvHeaderSaga } from './uv_header/uv_header.saga';
 
 const sagas = [uvHeaderSaga, uvPieSaga, uvBarChartSaga];
