+++ src/uv_bar-chart/uv_bar-chart.actions.ts
@@ -1,6 +1,6 @@
-import UV_BAR_CHART from './uv_bar_chart-constants';
-import UVBarChartConfig from './uv_bar_chart-interface-config';
-import UVItem from '../uv_interface-item';
+import UV_BAR_CHART from './uv_bar-chart.constants';
+import UVBarChartConfig from './uv_bar-chart.interface.config';
+import UVItem from '../uv_interface.item';
 
 const initBarChart = (barChartConfig: UVBarChartConfig, barChartItems: UVItem[]) => {
   return {
