+++ src/uv_bar-chart/uv_bar-chart.constants.ts
