+++ src/uv_bar-chart/uv_bar-chart.interface.config.ts
