+++ src/uv_header/uv_header.actions.ts
@@ -1,5 +1,5 @@
-import UV_HEADER from './uv_header-constants';
-import UVHeaderConfig from './uv_header-interface-config';
+import UV_HEADER from './uv_header.constants';
+import UVHeaderConfig from './uv_header.interface.config';
 
 const loadHeader = (actionConfig: UVHeaderConfig) => {
   return {
