+++ src/uv_header/uv_header.saga.ts
@@ -1,6 +1,6 @@
 import { takeEvery } from 'redux-saga/effects';
-import UVAction from '../uv_interface-action';
-import UV_HEADER from './uv_header-constants';
+import UVAction from '../uv_interface.action';
+import UV_HEADER from './uv_header.constants';
 
 function* updateHeader(pieSlice: UVAction) {
   // Note: We can dispatch event using store dispatch but it's not recommended in generator by Redux Saga docs.
