+++ src/uv_pie/uv_pie.actions.ts
@@ -1,6 +1,6 @@
-import UV_PIE from './uv_pie-constants';
-import UVPieConfig from './uv_pie-interface-config';
-import UVCategory from '../uv_interface-category';
+import UV_PIE from './uv_pie.constants';
+import UVPieConfig from './uv_pie.interface.config';
+import UVCategory from '../uv_interface.category';
 
 const loadPie = (pieConfig: UVPieConfig, pieCategories: UVCategory[]) => {
   return {
