+++ src/uv_pie/uv_pie.saga.ts
@@ -1,6 +1,6 @@
 import { takeEvery } from 'redux-saga/effects';
-import UV_PIE from './uv_pie-constants';
-import UVAction from '../uv_interface-action';
+import UV_PIE from './uv_pie.constants';
+import UVAction from '../uv_interface.action';
 
 function* updatePie(pieSlice: UVAction) {
   // Note: We can dispatch event using store dispatch but it's not recommended in generator by Redux Saga docs.
