+++ src/uv_pie/uv_pie.tsx
@@ -11,10 +11,10 @@ import am4themes_material from "@amcharts/amcharts4/themes/material";
 import './uv_pie.css';
 import { uvStore } from '../uv_store';
 import { useSelector } from 'react-redux';
-import { UVRootState } from '../root-reducer';
-import UVCategory from '../uv_interface-category';
-import UVAmount from '../uv_interface-amount';
-import { updateBarChart } from '../uv_bar_chart/uv_bar_chart-actions';
+import { UVRootState } from '../root.reducer';
+import UVCategory from '../uv_interface.category';
+import UVAmount from '../uv_interface.amount';
+import { updateBarChart } from '../uv_bar-chart/uv_bar-chart.actions';
 
 am4core.useTheme(am4themes_material);
 am4core.useTheme(am4themes_animated);
