+++ src/uv_store.ts
@@ -5,14 +5,14 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import * as appData from './uv_app-data.json';
 import * as headerData from './uv_header/uv_header.json';
 import * as pieData from './uv_pie/uv_pie.json';
-import * as barChartData from './uv_bar_chart/uv_bar_chart.json';
+import * as barChartData from './uv_bar-chart/uv_bar-chart.json';
 
-import { rootReducer } from './root-reducer';
-import { runAllSaga } from './root-sagas';
+import { rootReducer } from './root.reducer';
+import { runAllSaga } from './root.sagas';
 
-import { loadHeader } from './uv_header/uv_header-actions';
-import { loadPie } from './uv_pie/uv_pie-actions';
-import { initBarChart } from './uv_bar_chart/uv_bar_chart-actions';
+import { loadHeader } from './uv_header/uv_header.actions';
+import { loadPie } from './uv_pie/uv_pie.actions';
+import { initBarChart } from './uv_bar-chart/uv_bar-chart.actions';
 import uvObject from '@uv-tech/util/lib/uv-object';
 
 const sagaMiddleware = createSagaMiddleware();
